requirements.txt
discord.py
asyncio
json
requests
os
random